Agrego funcionalidad eliminar cuenta

+++ frontend/user-profile.js
@@ -1,4 +1,7 @@
 
+//  import { clear } from "console";
+//  import { use } from "../backend/src/routes/users";
+
 const nav = document.querySelector(".nav");
 const searchIcon = document.querySelector("#searchIcon");
 const navOpenBtn = document.querySelector(".navOpenBtn");
@@ -42,11 +45,61 @@ if(user){
         email.innerText = data.email;
     });
 }
-
+console.log(user)
 // Logout de usuario
-const logoutButton = document.getElementById('logout');
-logoutButton.addEventListener('click', () => {
-    // Limpiar localStorage
-    localStorage.clear();
-    window.location.href = 'inicio.html';
+
+const logoutButton = document.getElementById("logout");
+
+logoutButton.addEventListener('click', async () => {
+  try {
+    const userId = user.id;
+
+    const response = await fetch('http://localhost:3000/api/v1/logout', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ userId }), // Envía el ID del usuario al servidor
+    });
+
+    const data = await response.json();
+    if (response.ok) {
+      console.log(data.message);
+      localStorage.clear(); // Borra el almacenamiento local
+      window.location.href = 'inicio.html'; // Redirige al usuario a inicio
+    } else {
+      console.error('Error:', data.message);
+    }
+  } catch (error) {
+    console.error('Error al desloguear:', error);
+  }
+});
+
+// Eliminar usuario
+
+const deleteButton = document.getElementById('deleteUser');
+
+deleteButton.addEventListener('click', async () => {
+  const userId = user.id
+
+  const confirmation = confirm('¿Estás seguro de que deseas eliminar tu cuenta?');
+  if (!confirmation) return;
+
+  try {
+    const response = await fetch(`http://localhost:3000/api/v1/users/${userId}`, {
+      method: 'DELETE',
+    });
+
+    if (response.ok) {
+      alert('Usuario eliminado exitosamente');
+      localStorage.clear();
+      window.location.href = 'inicio.html';
+    } else {
+      const error = await response.json();
+      alert('Error al eliminar el usuario: ' + error.error);
+    }
+  } catch (err) {
+    console.error(err);
+    alert('Hubo un problema con la solicitud');
+  }
 });
